move location search function to util
